Revert "Put Items into its own category instead of being in settings"
This reverts commit 1188e467
